refactoring osdconfig/examples
