+++ osdconfig/examples/main/main.go
